Fix some compiler warnings

+++ audioengine.cpp
@@ -148,7 +148,7 @@ void AudioEngine::write(int16_t *pcm, size_t s)
 	m_outdev->write((const char *) pcm, sizeof(int16_t) * s);
 }
 
-uint16_t AudioEngine::read(int16_t *pcm, size_t s)
+uint16_t AudioEngine::read(int16_t *pcm, int s)
 {
 	if(m_audioinq.size() >= s){
 		for(int i = 0; i < s; ++i){
